+++ changelog.md
@@ -1,3 +1,7 @@
+# 1.5.1
+* Fixed incompat with control craft
+* Changed schematic renderer
+# 1.5.0
 * Added a way to open client/server setting via commands ("/vmod-client open-client-settings", "/vmod-client open-server-settings")
 * Added HUD Info Window if you hold toolgun and press "I"
 * Fixed issue when placing schematic multiple times